refactor: improve code quality, performance, and maintainability
- Fix OAuth callback JWT decoding with proper validation
- Add ErrorBoundary component for protected routes
- Split ProfileForm into smaller components (SocialInput, LocationSection, SocialAccountsSection)
- Extract error handling utility (getErrorMessage)
- Add React.memo to SessionRow for performance
- Create useDialogState hook for reusable dialog state management
- Lazy load react-image-crop in upload-avatar for smaller bundle
- Consolidate OAuth buttons into single configurable component
- Standardize button loading states using loading/loadingText props
