+++ main-app/src/app/(auth)/components/oauth-buttons.tsx
@@ -1,76 +1,80 @@
-"use client";
+'use client';
 
-import { Badge } from "@/components/ui/badge";
-import { useLastUsedProvider } from "@/hooks/auth/use-last-used-provider";
-import { useSafeRedirect } from "@/hooks/auth/use-safe-redirect";
+import { Badge } from '@/components/ui/badge';
+import { Button } from '@/components/ui/button';
+import { useLastUsedProvider } from '@/hooks/auth/use-last-used-provider';
+import { useSafeRedirect } from '@/hooks/auth/use-safe-redirect';
+import { GithubOAuthClickFunction, GoogleOAuthClickFunction } from '@/lib/apis';
+import { AuthMethod } from '@/types/user';
+import { useState, useCallback } from 'react';
+import { FaGithub, FaGoogle } from 'react-icons/fa6';
+import type { IconType } from 'react-icons';
 
-import { Button } from "@/components/ui/button";
-import { GithubOAuthClickFunction, GoogleOAuthClickFunction } from "@/lib/apis";
-import { AuthMethod } from "@/types/user";
-import { useState } from "react";
-import { FaGithub, FaGoogle } from "react-icons/fa6";
+type OAuthProvider = 'google' | 'github';
+
+interface OAuthConfig {
+  icon: IconType;
+  label: string;
+  authMethod: AuthMethod;
+  onClick: (options: { redirect?: string }) => void;
+}
+
+const oauthConfig: Record<OAuthProvider, OAuthConfig> = {
+  google: {
+    icon: FaGoogle,
+    label: 'Sign in with Google',
+    authMethod: AuthMethod.GOOGLE,
+    onClick: GoogleOAuthClickFunction,
+  },
+  github: {
+    icon: FaGithub,
+    label: 'Sign in with GitHub',
+    authMethod: AuthMethod.GITHUB,
+    onClick: GithubOAuthClickFunction,
+  },
+};
 
 interface OAuthButtonProps {
+  provider: OAuthProvider;
   redirectTo: string;
 }
 
-export const OAuthWithGoogle = ({ redirectTo }: OAuthButtonProps) => {
+function OAuthButton({ provider, redirectTo }: OAuthButtonProps) {
   const [isLoading, setIsLoading] = useState(false);
   const { isLastUsed } = useLastUsedProvider();
+  const config = oauthConfig[provider];
+  const Icon = config.icon;
 
-  const handleGoogleOAuth = () => {
+  const handleClick = useCallback(() => {
     setIsLoading(true);
-    GoogleOAuthClickFunction({ redirect: redirectTo });
-  };
+    config.onClick({ redirect: redirectTo });
+  }, [config, redirectTo]);
 
   return (
-    <Button
-      variant="outline"
-      className="relative"
-      onClick={handleGoogleOAuth}
-      disabled={isLoading}
-    >
-      <FaGoogle className="h-4 w-4" />
-      <span className="">Sign in with Google</span>
-      {isLastUsed(AuthMethod.GOOGLE) && (
-        <Badge className="absolute right-1 top-1">Last Used</Badge>
-      )}
+    <Button variant="outline" className="relative" onClick={handleClick} disabled={isLoading}>
+      <Icon className="h-4 w-4" />
+      <span>{config.label}</span>
+      {isLastUsed(config.authMethod) && <Badge className="absolute right-1 top-1">Last Used</Badge>}
     </Button>
   );
-};
+}
 
-export const OAuthWithGithub = ({ redirectTo }: OAuthButtonProps) => {
-  const [isLoading, setIsLoading] = useState(false);
-  const { isLastUsed } = useLastUsedProvider();
+// Keep individual exports for backward compatibility
+export const OAuthWithGoogle = ({ redirectTo }: { redirectTo: string }) => (
+  <OAuthButton provider="google" redirectTo={redirectTo} />
+);
 
-  const handleGithubOAuth = () => {
-    setIsLoading(true);
-    GithubOAuthClickFunction({ redirect: redirectTo });
-  };
-
-  return (
-    <Button
-      variant="outline"
-      className="relative"
-      onClick={handleGithubOAuth}
-      disabled={isLoading}
-    >
-      <FaGithub className="h-4 w-4" />
-      <span className="">Sign in with GitHub</span>
-      {isLastUsed(AuthMethod.GITHUB) && (
-        <Badge className="absolute right-1 top-1">Last Used</Badge>
-      )}
-    </Button>
-  );
-};
+export const OAuthWithGithub = ({ redirectTo }: { redirectTo: string }) => (
+  <OAuthButton provider="github" redirectTo={redirectTo} />
+);
 
 export const OAuthButtons = () => {
-  const redirectTo = useSafeRedirect("/dashboard");
+  const redirectTo = useSafeRedirect('/dashboard');
 
   return (
     <div className="flex flex-col gap-2">
-      <OAuthWithGoogle redirectTo={redirectTo} />
-      <OAuthWithGithub redirectTo={redirectTo} />
+      <OAuthButton provider="google" redirectTo={redirectTo} />
+      <OAuthButton provider="github" redirectTo={redirectTo} />
     </div>
   );
 };
