+++ main-app/src/components/error-boundary.tsx
@@ -1,0 +1,70 @@
+"use client";
+
+import { Button } from "@/components/ui/button";
+import { Component, type ReactNode } from "react";
+
+interface ErrorBoundaryProps {
+  children: ReactNode;
+  fallback?: ReactNode;
+}
+
+interface ErrorBoundaryState {
+  hasError: boolean;
+  error: Error | null;
+}
+
+export class ErrorBoundary extends Component<
+  ErrorBoundaryProps,
+  ErrorBoundaryState
+> {
+  constructor(props: ErrorBoundaryProps) {
+    super(props);
+    this.state = { hasError: false, error: null };
+  }
+
+  static getDerivedStateFromError(error: Error): ErrorBoundaryState {
+    return { hasError: true, error };
+  }
+
+  componentDidCatch(error: Error, errorInfo: React.ErrorInfo) {
+    console.error("Error caught by boundary:", error, errorInfo);
+  }
+
+  handleRetry = () => {
+    this.setState({ hasError: false, error: null });
+  };
+
+  render() {
+    if (this.state.hasError) {
+      if (this.props.fallback) {
+        return this.props.fallback;
+      }
+
+      return (
+        <div className="flex flex-col items-center justify-center min-h-[50vh] p-8">
+          <div className="text-center space-y-4 max-w-md">
+            <h2 className="text-2xl font-semibold text-foreground">
+              Something went wrong
+            </h2>
+            <p className="text-muted-foreground">
+              An unexpected error occurred. Please try again.
+            </p>
+            <div className="flex gap-3 justify-center">
+              <Button onClick={this.handleRetry} variant="default">
+                Try again
+              </Button>
+              <Button
+                onClick={() => (window.location.href = "/dashboard")}
+                variant="outline"
+              >
+                Go to Dashboard
+              </Button>
+            </div>
+          </div>
+        </div>
+      );
+    }
+
+    return this.props.children;
+  }
+}
