+++ main-app/src/lib/utils/get-error-message.ts
@@ -1,0 +1,23 @@
+import { AxiosError } from "axios";
+
+interface ApiErrorResponse {
+  message?: string;
+  error?: string;
+}
+
+export function getErrorMessage(error: unknown): string {
+  if (error instanceof AxiosError) {
+    const data = error.response?.data as ApiErrorResponse | undefined;
+    return data?.message || data?.error || error.message || "Request failed";
+  }
+
+  if (error instanceof Error) {
+    return error.message;
+  }
+
+  if (typeof error === "string") {
+    return error;
+  }
+
+  return "Something went wrong";
+}
