+++ marketing/src/app/layout.tsx
@@ -1,5 +1,4 @@
 import AppRootProvider from '@/providers/app-root-provider';
-import { RootProvider } from 'fumadocs-ui/provider/next';
 import { metadata } from '@/seo/metadata';
 import './globals.css';
 export { metadata };
@@ -12,9 +11,7 @@ export default function RootLayout({
   return (
     <html lang="en" suppressHydrationWarning>
       <body suppressHydrationWarning>
-        <AppRootProvider>
-          <RootProvider>{children}</RootProvider>
-        </AppRootProvider>
+        <AppRootProvider>{children}</AppRootProvider>
       </body>
     </html>
   );
